+++ src/cipher/zk.rs
@@ -111,38 +111,35 @@ pub fn decrypt(
 #[cfg(test)]
 mod tests {
     use crate::cipher::{decrypt, encrypt, PoseidonCipher};
-    use anyhow::Result;
     use dusk_bls12_381::BlsScalar;
     use dusk_jubjub::{dhke, JubJubExtended, GENERATOR_EXTENDED};
     use dusk_plonk::constraint_system::ecc::scalar_mul::variable_base::variable_base_scalar_mul;
-    use dusk_plonk::constraint_system::ecc::Point;
     use dusk_plonk::prelude::*;
+    use rand_core::OsRng;
 
     #[test]
-    fn gadget() -> Result<()> {
-        let mut rng = rand::thread_rng();
-
+    fn gadget() -> Result<(), Error> {
         // Generate a secret and a public key for Bob
-        let bob_secret = JubJubScalar::random(&mut rng);
+        let bob_secret = JubJubScalar::random(&mut OsRng);
 
         // Generate a secret and a public key for Alice
-        let alice_secret = JubJubScalar::random(&mut rng);
+        let alice_secret = JubJubScalar::random(&mut OsRng);
         let alice_public = GENERATOR_EXTENDED * &alice_secret;
 
         // Generate a shared secret
         let shared_secret = dhke(&bob_secret, &alice_public);
 
         // Generate a secret message
-        let a = BlsScalar::random(&mut rng);
-        let b = BlsScalar::random(&mut rng);
+        let a = BlsScalar::random(&mut OsRng);
+        let b = BlsScalar::random(&mut OsRng);
         let message = [a, b];
 
         // Perform the encryption
-        let nonce = BlsScalar::random(&mut rng);
+        let nonce = BlsScalar::random(&mut OsRng);
         let cipher = PoseidonCipher::encrypt(&message, &shared_secret, &nonce);
 
         let size = 13;
-        let pp = PublicParameters::setup(1 << size, &mut rng)?;
+        let pp = PublicParameters::setup(1 << size, &mut OsRng)?;
         let (ck, vk) = pp.trim(1 << size)?;
 
         let label = b"poseidon-cipher";
@@ -158,7 +155,7 @@ mod tests {
             let nonce = composer.add_input(nonce);
 
             let secret = composer.add_input((secret).into());
-            let public = Point::from_private_affine(composer, public.into());
+            let public = composer.add_affine(public.into());
 
             let shared = variable_base_scalar_mul(composer, secret, public);
 
@@ -170,7 +167,7 @@ mod tests {
             );
 
             let cipher_gadget =
-                encrypt(composer, shared.point(), nonce, &message_circuit);
+                encrypt(composer, &shared, nonce, &message_circuit);
 
             cipher.iter().zip(cipher_gadget.iter()).for_each(|(c, g)| {
                 let x = composer.add_input(*c);
@@ -178,7 +175,7 @@ mod tests {
             });
 
             let message_gadget =
-                decrypt(composer, shared.point(), nonce, &cipher_gadget);
+                decrypt(composer, &shared, nonce, &cipher_gadget);
 
             message
                 .iter()
@@ -213,7 +210,7 @@ mod tests {
         );
         verifier.preprocess(&ck)?;
 
-        assert!(verifier.verify(&proof, &vk, &vec![]).is_ok());
+        assert!(verifier.verify(&proof, &vk, &[BlsScalar::zero()]).is_ok());
 
         Ok(())
     }